fixing requested review changes
